HW17

+++ AQA-Makhortov/DTO/DapperTestsDTO/ProductDTO.cs
@@ -8,7 +8,7 @@
 
         string description,
 
-        long price,
+        double price,
 
         long stock,
